Enum aws_sdk_frauddetector::Error

#[non_exhaustive]
pub enum Error {
    AccessDeniedException(AccessDeniedException),
    ConflictException(ConflictException),
    InternalServerException(InternalServerException),
    ResourceNotFoundException(ResourceNotFoundException),
    ResourceUnavailableException(ResourceUnavailableException),
    ThrottlingException(ThrottlingException),
    ValidationException(ValidationException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

An exception indicating Amazon Fraud Detector does not have the needed permissions. This can occur if you submit a request, such as PutExternalModel, that specifies a role that is not in your account.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

ConflictException(ConflictException)

An exception indicating there was a conflict during a delete operation.

Tuple Fields of ConflictException

0: ConflictException

InternalServerException(InternalServerException)

An exception indicating an internal server error.

Tuple Fields of InternalServerException

0: InternalServerException

ResourceNotFoundException(ResourceNotFoundException)

An exception indicating the specified resource was not found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ResourceUnavailableException(ResourceUnavailableException)

An exception indicating that the attached customer-owned (external) model threw an exception when Amazon Fraud Detector invoked the model.

Tuple Fields of ResourceUnavailableException

0: ResourceUnavailableException

ThrottlingException(ThrottlingException)

An exception indicating a throttling error.

Tuple Fields of ThrottlingException

0: ThrottlingException

ValidationException(ValidationException)

An exception indicating a specified value is not allowed.

Tuple Fields of ValidationException

0: ValidationException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>