Module aws_sdk_frauddetector::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

resource_unavailable_exception

See ResourceUnavailableException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

An exception indicating Amazon Fraud Detector does not have the needed permissions. This can occur if you submit a request, such as PutExternalModel, that specifies a role that is not in your account.

BatchCreateVariableError

BatchGetVariableError

CancelBatchPredictionJobError

ConflictException

An exception indicating there was a conflict during a delete operation. The following delete operations can cause a conflict exception:

CreateBatchPredictionJobError

CreateDetectorVersionError

CreateModelError

CreateModelVersionError

CreateRuleError

CreateVariableError

DeleteBatchPredictionJobError

DeleteDetectorError

DeleteDetectorVersionError

DeleteEntityTypeError

DeleteEventError

DeleteEventTypeError

DeleteExternalModelError

DeleteLabelError

DeleteModelError

DeleteModelVersionError

DeleteOutcomeError

DeleteRuleError

DeleteVariableError

DescribeDetectorError

DescribeModelVersionsError

GetBatchPredictionJobsError

GetDetectorVersionError

GetDetectorsError

GetEntityTypesError

GetEventPredictionError

GetEventTypesError

GetExternalModelsError

GetKMSEncryptionKeyError

GetLabelsError

GetModelVersionError

GetModelsError

GetOutcomesError

GetRulesError

GetVariablesError

InternalServerException

An exception indicating an internal server error.

ListTagsForResourceError

PutDetectorError

PutEntityTypeError

PutEventTypeError

PutExternalModelError

PutKMSEncryptionKeyError

PutLabelError

PutOutcomeError

ResourceNotFoundException

An exception indicating the specified resource was not found. This can occur if you submit a request, such as CreateBatchPredictionJob, but the detector name or version does not exist.

ResourceUnavailableException

TagResourceError

ThrottlingException

An exception indicating a throttling error.

UntagResourceError

UpdateDetectorVersionError

UpdateDetectorVersionMetadataError

UpdateDetectorVersionStatusError

UpdateModelError

UpdateModelVersionError

UpdateModelVersionStatusError

UpdateRuleMetadataError

UpdateRuleVersionError

UpdateVariableError

ValidationException

An exception indicating a specified value is not allowed.

Enums

BatchCreateVariableErrorKind

BatchGetVariableErrorKind

CancelBatchPredictionJobErrorKind

CreateBatchPredictionJobErrorKind

CreateDetectorVersionErrorKind

CreateModelErrorKind

CreateModelVersionErrorKind

CreateRuleErrorKind

CreateVariableErrorKind

DeleteBatchPredictionJobErrorKind

DeleteDetectorErrorKind

DeleteDetectorVersionErrorKind

DeleteEntityTypeErrorKind

DeleteEventErrorKind

DeleteEventTypeErrorKind

DeleteExternalModelErrorKind

DeleteLabelErrorKind

DeleteModelErrorKind

DeleteModelVersionErrorKind

DeleteOutcomeErrorKind

DeleteRuleErrorKind

DeleteVariableErrorKind

DescribeDetectorErrorKind

DescribeModelVersionsErrorKind

GetBatchPredictionJobsErrorKind

GetDetectorVersionErrorKind

GetDetectorsErrorKind

GetEntityTypesErrorKind

GetEventPredictionErrorKind

GetEventTypesErrorKind

GetExternalModelsErrorKind

GetKMSEncryptionKeyErrorKind

GetLabelsErrorKind

GetModelVersionErrorKind

GetModelsErrorKind

GetOutcomesErrorKind

GetRulesErrorKind

GetVariablesErrorKind

ListTagsForResourceErrorKind

PutDetectorErrorKind

PutEntityTypeErrorKind

PutEventTypeErrorKind

PutExternalModelErrorKind

PutKMSEncryptionKeyErrorKind

PutLabelErrorKind

PutOutcomeErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDetectorVersionErrorKind

UpdateDetectorVersionMetadataErrorKind

UpdateDetectorVersionStatusErrorKind

UpdateModelErrorKind

UpdateModelVersionErrorKind

UpdateModelVersionStatusErrorKind

UpdateRuleMetadataErrorKind

UpdateRuleVersionErrorKind

UpdateVariableErrorKind