Module aws_sdk_frauddetector::model

Modules

batch_create_variable_error

See BatchCreateVariableError

batch_get_variable_error

See BatchGetVariableError

batch_prediction

See BatchPrediction

data_validation_metrics

See DataValidationMetrics

detector

See Detector

detector_version_summary

See DetectorVersionSummary

entity

See Entity

entity_type

See EntityType

event_type

See EventType

external_events_detail

See ExternalEventsDetail

external_model

See ExternalModel

external_model_outputs

See ExternalModelOutputs

external_model_summary

See ExternalModelSummary

field_validation_message

See FieldValidationMessage

file_validation_message

See FileValidationMessage

kms_key

See KmsKey

label

See Label

label_schema

See LabelSchema

log_odds_metric

See LogOddsMetric

metric_data_point

See MetricDataPoint

model

See Model

model_endpoint_data_blob

See ModelEndpointDataBlob

model_input_configuration

See ModelInputConfiguration

model_output_configuration

See ModelOutputConfiguration

model_scores

See ModelScores

model_version

See ModelVersion

model_version_detail

See ModelVersionDetail

outcome

See Outcome

rule

See Rule

rule_detail

See RuleDetail

rule_result

See RuleResult

tag

See Tag

training_data_schema

See TrainingDataSchema

training_metrics

See TrainingMetrics

training_result

See TrainingResult

variable

See Variable

variable_entry

See VariableEntry

variable_importance_metrics

See VariableImportanceMetrics

Structs

BatchCreateVariableError

Provides the error of the batch create variable API.

BatchGetVariableError

Provides the error of the batch get variable API.

BatchPrediction

The batch prediction details.

DataValidationMetrics

The model training validation messages.

Detector

The detector.

DetectorVersionSummary

The summary of the detector version.

Entity

The entity details.

EntityType

The entity type details.

EventType

The event type details.

ExternalEventsDetail

Details for the external events data used for model version training.

ExternalModel

The Amazon SageMaker model.

ExternalModelOutputs

The fraud prediction scores from Amazon SageMaker model.

ExternalModelSummary

The Amazon SageMaker model.

FieldValidationMessage

The message details.

FileValidationMessage

The message details.

KmsKey

The KMS key details.

Label

The label details.

LabelSchema

The label schema.

LogOddsMetric

The log odds metric details.

MetricDataPoint

Model performance metrics data points.

Model

The model.

ModelEndpointDataBlob

A pre-formed Amazon SageMaker model input you can include if your detector version includes an imported Amazon SageMaker model endpoint with pass-through input configuration.

ModelInputConfiguration

The Amazon SageMaker model input configuration.

ModelOutputConfiguration

Provides the Amazon Sagemaker model output configuration.

ModelScores

The fraud prediction scores.

ModelVersion

The model version.

ModelVersionDetail

The details of the model version.

Outcome

The outcome.

Rule

A rule.

RuleDetail

The details of the rule.

RuleResult

The rule results.

Tag

A key and value pair.

TrainingDataSchema

The training data schema.

TrainingMetrics

The training metric details.

TrainingResult

The training result details.

Variable

The variable.

VariableEntry

A variable in the list of variables for the batch create variable request.

VariableImportanceMetrics

The variable importance metrics details.

Enums

AsyncJobStatus

DataSource

DataType

DetectorVersionStatus

Language

ModelEndpointStatus

ModelInputDataFormat

ModelOutputDataFormat

ModelSource

ModelTypeEnum

ModelVersionStatus

RuleExecutionMode

TrainingDataSourceEnum