Module aws_sdk_frauddetector::operation

Structs

BatchCreateVariable

Creates a batch of variables.

BatchGetVariable

Gets a batch of variables.

CancelBatchPredictionJob

Cancels the specified batch prediction job.

CreateBatchPredictionJob

Creates a batch prediction job.

CreateDetectorVersion

Creates a detector version. The detector version starts in a DRAFT status.

CreateModel

Creates a model using the specified model type.

CreateModelVersion

Creates a version of the model using the specified model type and model id.

CreateRule

Creates a rule for use with the specified detector.

CreateVariable

Creates a variable.

DeleteBatchPredictionJob

Deletes a batch prediction job.

DeleteDetector

Deletes the detector. Before deleting a detector, you must first delete all detector versions and rule versions associated with the detector.

DeleteDetectorVersion

Deletes the detector version. You cannot delete detector versions that are in ACTIVE status.

DeleteEntityType

Deletes an entity type.

DeleteEvent

Deletes the specified event.

DeleteEventType

Deletes an event type.

DeleteExternalModel

Removes a SageMaker model from Amazon Fraud Detector.

DeleteLabel

Deletes a label.

DeleteModel

Deletes a model.

DeleteModelVersion

Deletes a model version.

DeleteOutcome

Deletes an outcome.

DeleteRule

Deletes the rule. You cannot delete a rule if it is used by an ACTIVE or INACTIVE detector version.

DeleteVariable

Deletes a variable.

DescribeDetector

Gets all versions for a specified detector.

DescribeModelVersions

Gets all of the model versions for the specified model type or for the specified model type and model ID. You can also get details for a single, specified model version.

GetBatchPredictionJobs

Gets all batch prediction jobs or a specific job if you specify a job ID. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 50 records per page. If you provide a maxResults, the value must be between 1 and 50. To get the next page results, provide the pagination token from the GetBatchPredictionJobsResponse as part of your request. A null pagination token fetches the records from the beginning.

GetDetectorVersion

Gets a particular detector version.

GetDetectors

Gets all detectors or a single detector if a detectorId is specified. This is a paginated API. If you

GetEntityTypes

Gets all entity types or a specific entity type if a name is specified. This is a paginated API. If you

GetEventPrediction

Evaluates an event against a detector version. If a version ID is not provided, the detector’s (ACTIVE) version is used.

GetEventTypes

Gets all event types or a specific event type if name is provided. This is a paginated API. If you

GetExternalModels

Gets the details for one or more Amazon SageMaker models that have been imported into the

GetKMSEncryptionKey

Gets the encryption key if a Key Management Service (KMS) customer master key (CMK) has been specified to be used to encrypt content in Amazon Fraud Detector.

GetLabels

Gets all labels or a specific label if name is provided. This is a paginated API. If you

GetModelVersion

Gets the details of the specified model version.

GetModels

Gets one or more models. Gets all models for the Amazon Web Services account if no model type and no model id provided. Gets all models for the Amazon Web Services account and model type, if the model type is specified but model id is not provided. Gets a specific model if (model type, model id) tuple is specified.

GetOutcomes

Gets one or more outcomes. This is a paginated

GetRules

Get all rules for a detector (paginated) if ruleId and ruleVersion are not specified. Gets all rules for the detector and the ruleId if present (paginated). Gets a specific rule if both the ruleId and the ruleVersion are specified.

GetVariables

Gets all of the variables or the specific variable. This is a

ListTagsForResource

Lists all tags associated with the resource. This is a paginated API. To get the next page results, provide the pagination token from the

PutDetector

Creates or updates a detector.

PutEntityType

Creates or updates an entity type. An entity represents who is performing the event. As part of a fraud prediction, you pass the entity ID to indicate the specific entity who performed the event. An entity type classifies the entity. Example classifications include customer, merchant, or account.

PutEventType

Creates or updates an event type. An event is a business activity that is evaluated for fraud risk. With Amazon Fraud Detector, you generate fraud predictions for events. An event type defines the structure for an event sent to Amazon Fraud Detector. This includes the variables sent as part of the event, the entity performing the event (such as a customer), and the labels that classify the event. Example event types include online payment transactions, account registrations, and authentications.

PutExternalModel

Creates or updates an Amazon SageMaker model endpoint. You can also use this action to update the configuration of the model endpoint, including the IAM role and/or the mapped variables.

PutKMSEncryptionKey

Specifies the Key Management Service (KMS) customer master key (CMK) to be used to encrypt content in Amazon Fraud Detector.

PutLabel

Creates or updates label. A label classifies an event as fraudulent or legitimate. Labels are associated with event types and used to train supervised machine learning models in Amazon Fraud Detector.

PutOutcome

Creates or updates an outcome.

TagResource

Assigns tags to a resource.

UntagResource

Removes tags from a resource.

UpdateDetectorVersion

Updates a detector version. The detector version attributes that you can update include models, external model endpoints, rules, rule execution mode, and description. You can only update a DRAFT detector version.

UpdateDetectorVersionMetadata

Updates the detector version's description. You can update the metadata for any detector version (DRAFT, ACTIVE, or

UpdateDetectorVersionStatus

Updates the detector version’s status. You can perform the following promotions or

UpdateModel

Updates a model. You can update the description attribute using this action.

UpdateModelVersion

Updates a model version. Updating a model version retrains an existing model version using updated training data and produces a new minor version of the model. You can update the training data set location and data access role attributes using this action. This action creates and trains a new minor version of the model, for example version 1.01, 1.02, 1.03.

UpdateModelVersionStatus

Updates the status of a model version.

UpdateRuleMetadata

Updates a rule's metadata. The description attribute can be updated.

UpdateRuleVersion

Updates a rule version resulting in a new rule version. Updates a rule version resulting in a new rule version (version 1, 2, 3 ...).

UpdateVariable

Updates a variable.